Organize components to subdirectories and code cleanup
