+++ frontend/src/components/album/AllAlbumsCard.jsx
