+++ frontend/src/components/core/Navbar.jsx
