+++ frontend/src/components/event/EventCard.jsx
