+++ frontend/src/components/opportunity/OpportunityCard.jsx
