+++ frontend/src/components/opportunity/OpportunityPane.jsx
