+++ frontend/src/components/post/NewPostCard.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPlus } from "@fortawesome/free-solid-svg-icons";
 import { UisArrowUpRight } from "@iconscout/react-unicons-solid";
