+++ frontend/src/components/post/NewPostCardUploadPhoto.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import { useState } from "react";
 
 {/* @author Jude Angelo Ilumin*/}
