+++ frontend/src/components/search/UserSearchCard.jsx
@@ -1,0 +1,9 @@
+const UserSearchCard = ({user}) => {
+  return (
+    <>
+
+    </>
+  )
+}
+
+export default UserSearchCard
