+++ frontend/src/layouts/RootLayout.jsx
@@ -1,5 +1,4 @@
-import React from "react";
-import SideBar from "../components/Sidebar";
+import SideBar from "../components/core/Sidebar";
 import { Outlet } from "react-router-dom";
 import { useLocation } from "react-router-dom";
 import { useNavigate } from "react-router-dom";
